fix(onboarding): repair native CI gates
